Make ContextMockerTrait easier to use
